refactor: group related Auth components
