+++ frontend/src/components/Auth/AuthButton.vue
